[lab6] net : add tcp and tcp socket

+++ os/lab6/net/timer.c
@@ -8,7 +8,7 @@
 
 static LIST_HEAD(timers);
 struct spinlock lock;
-static int tick = 0;
+static uint64_t tick = 0;
 
 
 __attribute((constructor)) init_timer_lock()  {
